fix: do not delete inprogress log

+++ core/src/main/scala/org/apache/spark/deploy/history/FsHistoryProvider.scala
@@ -546,7 +546,7 @@ private[history] class FsHistoryProvider(conf: SparkConf, clock: Clock)
               // try to parse it. This will allow the cleaner code to detect the file as stale
               // later on if it was not possible to parse it.
               try {
-                if (conf.get(CLEANER_ENABLED) && reader.modificationTime <
+                if (conf.get(CLEANER_ENABLED) && reader.completed && reader.modificationTime <
                     clock.getTimeMillis() - conf.get(MAX_LOG_AGE_S) * 1000) {
                   logInfo(s"Deleting expired event log ${reader.rootPath.toString}")
                   deleteLog(fs, reader.rootPath)
